extract callback-related types;
