+++ bindings/ergo-lib-c-core/src/rest/api.rs
@@ -1,5 +1,6 @@
 //! REST API for the services in Ergo ecosystem (node, explorer, etc.)
 
+pub mod callback;
 pub mod node;
 pub mod node_async;
 pub mod request_handle;
